pom.xml: make service file inclusion in jar work...
But my WORD is it ugly.

This uses the maven antrun plugin to run a task _after_ the jar has been
generated; this task appends the generated service file to the jar.

I had no choice but to proceed like this... The shade plugin wouldn't work,
includes in the jar plugin neither... A nightmare.